fixes #1685 where ofGetElapsedTimef() and ofGetElapsedTimeMicros() wrap after ~71 minutes

+++ libs/openFrameworks/utils/ofUtils.h
@@ -14,8 +14,8 @@ int 	ofNextPow2(int input);
 
 void	ofResetElapsedTimeCounter();		// this happens on the first frame
 float 	ofGetElapsedTimef();
-unsigned long ofGetElapsedTimeMillis();
-unsigned long ofGetElapsedTimeMicros();
+unsigned long long ofGetElapsedTimeMillis();
+unsigned long long ofGetElapsedTimeMicros();
 int 	ofGetFrameNum();
 
 int 	ofGetSeconds();
@@ -25,8 +25,8 @@ int 	ofGetHours();
 //number of seconds since 1970
 unsigned int ofGetUnixTime();
 
-unsigned long ofGetSystemTime( );			// system time in milliseconds;
-unsigned long ofGetSystemTimeMicros( );			// system time in microseconds;
+unsigned long long ofGetSystemTime( );			// system time in milliseconds;
+unsigned long long ofGetSystemTimeMicros( );			// system time in microseconds;
 
 		//returns 
 string ofGetTimestampString();
